Refactor FileNodeListFragment: simpily parsing loop
this removes the structures which checks for the end of
the FileNodeListFragment. Moreove, it removes FileNode skipping,
since all types of Nodes can be parsed now. And finally, the
FileNodeListFragment's padding skipper is removed because its prone
to found invalid FileNodes when the padding is not all zerod.

Additionally a switch for TransactionEntry's count for specific
FileNodeListFragment is added, though, there is no direct way
to get that information yet. Maybe, the transaction log could
be parsed for each FileNodeListFragment if the overhead is
small, since it's location is given in the header also found
in the input stream which is given to FileNodeListFragment.